Add support for GET /{network}/block/{height}/header
Currently one must query the whole block in order to get the information
in the header. This commit mimics the existing endpoint to only get the
transactions for a block.

With this commit it is possible to query only the headers for blocks.
This will reduce both bandwidth and payload sizes for the one parsing
the result.