Remove redundant check.
The diff looks more substantial, but the only change is to remove the check
`if self.sync.is_gateway_mode()` and to reformat the code.

As discussed on Slack, and as can be seen examining the code, the test is always
true, since a validator is always in gateway mode.

Although not a big change, it's one less thing to see and remember.