Rename `BlockSyn` ledger component.
As discussed on Slack.

The `canon` name, and the comment that accompanies it, refer to an earlier
version in which there was a map derived from the ledger. But that was later
replaced with the whole ledger (service).

Besides the code, most of the comments were taken care of by RustRover's
renaming, but a few needed manual tweaking.

The changes are confined to a single file, as expected.